+++ CHANGELOG.rst
@@ -4,4 +4,12 @@ Changelog
 
 Unreleased
 ==========
+* Added grouper model
+* Added versioning config in cms_config
+* Added versioning factory
+* Updated versioning config in cms_config and created default copy method testcase.
+
+0.0.11
+==========
+* Fixed missing migration and initial migration dependencies on the cms
 * Removed Python 3.5 EOL and added Python 3.7 and 3.8 in the tox and CircleCI configs